pyproject.toml
[project]
name = "birbbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Ethan",email = "ethan.harriz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py (>=2.6.2,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
